navbar/footer logos + cleanup
